requirements.txt
boto3>=1.34,<2
oracledb>=2,<3
python-dotenv>=1,<2
